requirements.txt
alabaster==0.7.4
amqp==1.4.6
anyjson==0.3.3
astroid==1.3.8
awsebcli==3.6.2
Babel==1.3
babelfish==0.5.5
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.20
blessed==1.9.5
boto==2.38.0
botocore==1.3.15
caniusepython3==3.3.0
celery==3.1.18
cement==2.4.0
cffi==1.1.0
characteristic==14.3.0
chardet==2.3.0
colorama==0.3.3
cryptography==0.9
cssselect==0.9.1
diff-match-patch==20121119
distlib==0.2.1
Django==1.8.4
django-celery==3.1.16
django-cors-headers==1.1.0
django-debug-toolbar==1.3.0
-e git+https://github.com/maherpk/django-devserver-redux.git@a55bdb9e7fe3a4d17eaf1af3bc9799ac6503022b#egg=django_devserver_redux
django-extensions==1.6.1
django-grappelli==2.6.5
django-haystack==2.4.0
django-import-export==0.2.8
django-ipware==1.1.2
django-model-utils==2.2
django-money==0.6.0
django-pgjson==0.3.1
django-photologue==3.2
django-registration-redux==1.1
django-sortedm2m==0.10.0
django-storages==1.1.8
django-taggit==0.14.0
django-widget-tweaks==1.3
djangorestframework==3.1.3
djrill==1.4.0
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
drf-extensions==0.2.7
ecdsa==0.13
enum34==1.0.4
Fabric==1.10.1
factory-boy==2.6.0
fake-factory==0.5.3
futures==3.0.3
geoip2==2.2.0
gnureadline==6.3.3
googlemaps==2.4
guessit==1.0.2
guppy==0.1.10
hachoir-core==1.3.3
hachoir-metadata==1.3.3
hachoir-parser==1.3.4
httpagentparser==1.7.8
idna==2.0
ipaddr==2.1.11
ipaddress==1.0.7
ipython==3.1.0
Jinja2==2.8
jmespath==0.7.1
jsonfield==1.0.3
kombu==3.0.26
line-profiler==1.0
logilab-common==1.0.2
lxml==3.5.0
MarkupSafe==0.23
maxminddb==1.2.0
meld3==1.0.2
ndg-httpsclient==0.4.0
numpy==1.10.4
paramiko==1.16.0
pathspec==0.3.3
pbr==1.8.1
Pillow==2.8.1
pip-tools==0.3.6
psycopg2==2.6.1
py-moneyed==0.5.0
py2neo==2.0.8
pyasn1==0.1.7
pyasn1-modules==0.0.8
pycparser==2.12
pycrypto==2.6.1
Pygments==2.0.2
pylint==1.4.4
pyOpenSSL==0.15.1
python-dateutil==1.5
pytz==2015.4
PyYAML==3.11
queuelib==1.4.2
raven==5.3.1
reportlab==3.2.0
requests==2.9.0
Scrapy==1.0.3
scrapy-djangoitem==1.0.0
service-identity==14.0.0
shortuuid==0.4.2
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
sqlparse==0.1.15
stevedore==1.10.0
stripe==1.22.2
supervisor==3.2.0
tablib==0.10.0
texttable==0.8.4
tvdb-api==1.10
Twisted==15.5.0
unicodecsv==0.14.1
w3lib==1.13.0
wcwidth==0.1.5
websocket-client==0.32.0
Werkzeug==0.10.4
wheel==0.24.0
yapf==0.3.1
zope.interface==4.1.3
